+++ .babelrc
@@ -6,8 +6,6 @@
       }
     }]
   ],
-  "ignore": [
-    "node_modules"
-  ],
+  "ignore": [],
   "plugins": ["transform-flow-strip-types"]
 }
