+++ server/.babelrc
@@ -7,9 +7,7 @@
     }]
   ],
   "ignore": [
-    "node_modules",
-    "views/partials/analytics.js.njk",
-    "views/partials/hotjar.js.njk"
+    "nothings"
   ],
   "plugins": ["transform-flow-strip-types"]
 }
