+++ server/package.json
@@ -2,11 +2,9 @@
   "name": "common",
   "version": "0.0.0",
   "description": "The node server serving content for wellcomecollection.org",
-  "main": ".dist/index.js",
-  "module": "./index.js",
   "scripts": {
     "app:build": "../node_modules/.bin/webpack",
-    "app:dev": "../node_modules/.bin/nodemon -e js,njk,json --exec babel-node --inspect run.js",
+    "app:dev": "../node_modules/.bin/nodemon -e js,njk,json --exec ../node_modules/.bin/babel-node --inspect=4000 app.js",
     "optimise:svg": "node ./optimise-svgs.js",
     "test": "ava && flow",
     "test:accessibility": "pa11y",
@@ -23,16 +21,13 @@
   "license": "MIT",
   "dependencies": {
     "ava": "^0.17.0",
-    "babel-cli": "^6.26.0",
     "babel-loader": "^7.1.2",
-    "babel-plugin-transform-flow-strip-types": "^6.22.0",
-    "babel-preset-env": "^1.1.8",
     "chokidar": "^1.7.0",
     "copy-webpack-plugin": "^4.2.3",
     "entities": "^1.1.1",
     "flow-bin": "^0.57.3",
     "immutable": "^4.0.0-rc.9",
-    "interval-cache": "0.0.1",
+    "interval-cache": "1.0.0",
     "koa": "^2.4.1",
     "koa-router": "^7.1.0",
     "koa-static": "^3.0.0",
@@ -56,10 +51,10 @@
     "browserstack": "^1.5.0",
     "browserstack-local": "^1.3.0",
     "concurrently": "^3.1.0",
+    "flow-bro": "^0.4.1",
     "open": "0.0.5",
     "pa11y": "^4.12.1",
-    "svgo": "^0.7.2",
-    "flow-bro": "^0.4.1"
+    "svgo": "^0.7.2"
   },
   "ava": {
     "require": [
